refactor(Reservation): domain 패키지로 이동 #step3
