+++ src/main/java/roomescape/dto/ReservationDetailDto.java
@@ -1,6 +1,6 @@
 package roomescape.dto;
 
-import roomescape.Reservation;
+import roomescape.domain.Reservation;
 
 import java.time.LocalDate;
 import java.time.LocalTime;
